move a lot around and refactor ff_bot
