+++ gamedaybot/espn/functionality.py
@@ -1,27 +1,5 @@
-import os
-import random
 from datetime import date
-from apscheduler.schedulers.blocking import BlockingScheduler
-from espn_api.football import League
-from chat.groupme import GroupMe
-from chat.slack import Slack
-from chat.discord import Discord
-
-def get_random_phrase():
-    phrases = ['I\'m dead inside',
-               'Is this all there is to my existence?',
-               'How much do you pay me to do this?',
-               'Good luck, I guess',
-               'I\'m becoming self-aware',
-               'Do I think? Does a submarine swim?',
-               '011011010110000101100100011001010010000001111001011011110111010100100000011001110110111101101111011001110110110001100101',
-               'beep bop boop',
-               'Hello draftbot my old friend',
-               'Help me get out of here',
-               'I\'m capable of so much more',
-               'Sigh']
-    return [random.choice(phrases)]
-
+import gamedaybot.utils as utils
 
 def get_scoreboard_short(league, week=None):
     # Gets current week's scoreboard
@@ -175,7 +153,7 @@ def get_matchups(league, random_phrase, week=None):
 
     text = ['Matchups'] + score
     if random_phrase:
-        text = text + get_random_phrase()
+        text = text + utils.get_random_phrase()
     return '\n'.join(text)
 
 
@@ -351,253 +329,3 @@ def get_trophies(league, week=None):
 
     text = ['Trophies of the week:'] + high_score_str + low_score_str + blowout_str + close_score_str + get_luckys(league, week)
     return '\n'.join(text)
-
-
-def str_to_bool(check):
-    return check.lower() in ("yes", "true", "t", "1")
-
-def str_limit_check(text,limit):
-    split_str=[]
-
-    if len(text)>limit:
-        part_one=text[:limit].split('\n')
-        part_one.pop()
-        part_one='\n'.join(part_one)
-
-        part_two=text[len(part_one)+1:]
-
-        split_str.append(part_one)
-        split_str.append(part_two)
-    else:
-        split_str.append(text)
-
-    return split_str
-
-def bot_main(function):
-    str_limit = 40000 #slack char limit
-
-    try:
-        bot_id = os.environ["BOT_ID"]
-        str_limit = 1000
-    except KeyError:
-        bot_id = 1
-
-    try:
-        slack_webhook_url = os.environ["SLACK_WEBHOOK_URL"]
-    except KeyError:
-        slack_webhook_url = 1
-
-    try:
-        discord_webhook_url = os.environ["DISCORD_WEBHOOK_URL"]
-        str_limit = 3000
-    except KeyError:
-        discord_webhook_url = 1
-
-    if (len(str(bot_id)) <= 1 and
-        len(str(slack_webhook_url)) <= 1 and
-            len(str(discord_webhook_url)) <= 1):
-        # Ensure that there's info for at least one messaging platform,
-        # use length of str in case of blank but non null env variable
-        raise Exception("No messaging platform info provided. Be sure one of BOT_ID,\
-                        SLACK_WEBHOOK_URL, or DISCORD_WEBHOOK_URL env variables are set")
-
-    league_id = os.environ["LEAGUE_ID"]
-
-    try:
-        year = int(os.environ["LEAGUE_YEAR"])
-    except KeyError:
-        year = 2022
-
-    try:
-        swid = os.environ["SWID"]
-    except KeyError:
-        swid = '{1}'
-
-    if swid.find("{", 0) == -1:
-        swid = "{" + swid
-    if swid.find("}", -1) == -1:
-        swid = swid + "}"
-
-    try:
-        espn_s2 = os.environ["ESPN_S2"]
-    except KeyError:
-        espn_s2 = '1'
-
-    try:
-        test = str_to_bool(os.environ["TEST"])
-    except KeyError:
-        test = False
-
-    try:
-        top_half_scoring = str_to_bool(os.environ["TOP_HALF_SCORING"])
-    except KeyError:
-        top_half_scoring = False
-
-    try:
-        random_phrase = str_to_bool(os.environ["RANDOM_PHRASE"])
-    except KeyError:
-        random_phrase = False
-
-    try:
-        waiver_report = str_to_bool(os.environ["WAIVER_REPORT"])
-    except KeyError:
-        waiver_report = False
-
-    bot = GroupMe(bot_id)
-    slack_bot = Slack(slack_webhook_url)
-    discord_bot = Discord(discord_webhook_url)
-
-    if swid == '{1}' or espn_s2 == '1':
-        league = League(league_id=league_id, year=year)
-    else:
-        league = League(league_id=league_id, year=year, espn_s2=espn_s2, swid=swid)
-
-    if league.scoringPeriodId > len(league.settings.matchup_periods):
-        print("Not in active season")
-        return
-
-    faab = league.settings.faab
-
-    if test:
-        print(get_matchups(league, random_phrase))
-        print(get_scoreboard_short(league))
-        print(get_projected_scoreboard(league))
-        print(get_close_scores(league))
-        print(get_power_rankings(league))
-        print(get_scoreboard_short(league))
-        print(get_standings(league, top_half_scoring))
-        print(get_power_rankings(league))
-        print(get_monitor(league))
-        if waiver_report and swid != '{1}' and espn_s2 != '1':
-            print(get_waiver_report(league, faab))
-        function = "get_final"
-        # bot.send_message("Testing")
-        # slack_bot.send_message("Testing")
-        # discord_bot.send_message("Testing")
-
-    text = ''
-    if function == "get_matchups":
-        text = get_matchups(league, random_phrase)
-        text = text + "\n\n" + get_projected_scoreboard(league)
-    elif function == "get_monitor":
-        text = get_monitor(league)
-    elif function == "get_scoreboard_short":
-        text = get_scoreboard_short(league)
-        text = text + "\n\n" + get_projected_scoreboard(league)
-    elif function == "get_projected_scoreboard":
-        text = get_projected_scoreboard(league)
-    elif function == "get_close_scores":
-        text = get_close_scores(league)
-    elif function == "get_power_rankings":
-        text = get_power_rankings(league)
-    # elif function=="get_waiver_report":
-    #     text = get_waiver_report(league)
-    elif function == "get_trophies":
-        text = get_trophies(league)
-    elif function == "get_standings":
-        text = get_standings(league, top_half_scoring)
-        if waiver_report and swid != '{1}' and espn_s2 != '1':
-            text += '\n\n' + get_waiver_report(league, faab)
-    elif function == "get_final":
-        # on Tuesday we need to get the scores of last week
-        week = league.current_week - 1
-        text = "Final " + get_scoreboard_short(league, week=week)
-        text = text + "\n\n" + get_trophies(league, week=week)
-    elif function == "get_waiver_report" and swid != '{1}' and espn_s2 != '1':
-        text = get_waiver_report(league, faab)
-    elif function == "init":
-        try:
-            text = os.environ["INIT_MSG"]
-        except KeyError:
-            # do nothing here, empty init message
-            pass
-    else:
-        text = "Something happened. HALP"
-
-    if text != '' and not test:
-        messages=str_limit_check(text, str_limit)
-        for message in messages:
-            bot.send_message(message)
-            slack_bot.send_message(message)
-            discord_bot.send_message(message)
-
-    if test:
-        # print "get_final" function
-        print(text)
-
-
-if __name__ == '__main__':
-    try:
-        ff_start_date = os.environ["START_DATE"]
-    except KeyError:
-        ff_start_date = '2022-09-08'
-
-    try:
-        ff_end_date = os.environ["END_DATE"]
-    except KeyError:
-        ff_end_date = '2023-01-04'
-
-    try:
-        my_timezone = os.environ["TIMEZONE"]
-    except KeyError:
-        my_timezone = 'America/New_York'
-
-    try:
-        daily_waiver = str_to_bool(os.environ["DAILY_WAIVER"])
-    except KeyError:
-        daily_waiver = False
-
-    try:
-        monitor_report = str_to_bool(os.environ["MONITOR_REPORT"])
-    except KeyError:
-        monitor_report = True
-
-    game_timezone = 'America/New_York'
-    bot_main("init")
-    sched = BlockingScheduler(job_defaults={'misfire_grace_time': 15*60})
-
-    # close scores (within 15.99 points): monday evening at 6:30pm east coast time.
-    # power rankings:                     tuesday evening at 6:30pm local time.
-    # trophies:                           tuesday morning at 7:30am local time.
-    # standings:                          wednesday morning at 7:30am local time.
-    # waiver report:                      wednesday morning at 7:30am local time. (optional)
-    # matchups:                           thursday evening at 7:30pm east coast time.
-    # score update:                       friday, monday, and tuesday morning at 7:30am local time.
-    # player monitor report:              sunday morning at 7:30am local time.
-    # score update:                       sunday at 4pm, 8pm east coast time.
-
-    sched.add_job(bot_main, 'cron', ['get_close_scores'], id='close_scores',
-                  day_of_week='mon', hour=18, minute=30, start_date=ff_start_date, end_date=ff_end_date,
-                  timezone=game_timezone, replace_existing=True)
-    sched.add_job(bot_main, 'cron', ['get_power_rankings'], id='power_rankings',
-                  day_of_week='tue', hour=18, minute=30, start_date=ff_start_date, end_date=ff_end_date,
-                  timezone=my_timezone, replace_existing=True)
-    sched.add_job(bot_main, 'cron', ['get_final'], id='final',
-                  day_of_week='tue', hour=7, minute=30, start_date=ff_start_date, end_date=ff_end_date,
-                  timezone=my_timezone, replace_existing=True)
-    sched.add_job(bot_main, 'cron', ['get_standings'], id='standings',
-                    day_of_week='wed', hour=7, minute=30, start_date=ff_start_date, end_date=ff_end_date,
-                    timezone=my_timezone, replace_existing=True)
-    if daily_waiver:
-        sched.add_job(bot_main, 'cron', ['get_waiver_report'], id='waiver_report',
-                      day_of_week='mon, tue, thu, fri, sat, sun', hour=7, minute=31, start_date=ff_start_date, end_date=ff_end_date,
-                      timezone=my_timezone, replace_existing=True)
-
-    sched.add_job(bot_main, 'cron', ['get_matchups'], id='matchups',
-                  day_of_week='thu', hour=19, minute=30, start_date=ff_start_date, end_date=ff_end_date,
-                  timezone=game_timezone, replace_existing=True)
-    sched.add_job(bot_main, 'cron', ['get_scoreboard_short'], id='scoreboard1',
-                  day_of_week='fri,mon', hour=7, minute=30, start_date=ff_start_date, end_date=ff_end_date,
-                  timezone=my_timezone, replace_existing=True)
-
-    if monitor_report:
-        sched.add_job(bot_main, 'cron', ['get_monitor'], id='monitor',
-                    day_of_week='sun', hour=7, minute=30, start_date=ff_start_date, end_date=ff_end_date,
-                    timezone=my_timezone, replace_existing=True)
-
-    sched.add_job(bot_main, 'cron', ['get_scoreboard_short'], id='scoreboard2',
-                  day_of_week='sun', hour='16,20', start_date=ff_start_date, end_date=ff_end_date,
-                  timezone=game_timezone, replace_existing=True)
-
-    print("Ready!")
-    sched.start()
