+++ gamedaybot/utils.py
@@ -1,0 +1,36 @@
+import random
+
+def get_random_phrase():
+    phrases = ['I\'m dead inside',
+               'Is this all there is to my existence?',
+               'How much do you pay me to do this?',
+               'Good luck, I guess',
+               'I\'m becoming self-aware',
+               'Do I think? Does a submarine swim?',
+               '011011010110000101100100011001010010000001111001011011110111010100100000011001110110111101101111011001110110110001100101',
+               'beep bop boop',
+               'Hello draftbot my old friend',
+               'Help me get out of here',
+               'I\'m capable of so much more',
+               'Sigh']
+    return [random.choice(phrases)]
+
+def str_to_bool(check):
+    return check.lower() in ("yes", "true", "t", "1")
+
+def str_limit_check(text,limit):
+    split_str=[]
+
+    if len(text)>limit:
+        part_one=text[:limit].split('\n')
+        part_one.pop()
+        part_one='\n'.join(part_one)
+
+        part_two=text[len(part_one)+1:]
+
+        split_str.append(part_one)
+        split_str.append(part_two)
+    else:
+        split_str.append(text)
+
+    return split_str
